services/rss-agent/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
feedgen==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
boto3==1.34.0
apscheduler==3.10.4
httpx==0.25.2
lxml==4.9.3
celery==5.3.4
redis==5.0.1
flower==2.0.1
tenacity==8.2.3
aiofiles==23.2.1
python-dotenv==1.0.0